requirements.txt
aiofiles==0.6.0
aioredis==1.3.1
async-timeout==3.0.1
Cerberus==1.3.2
certifi==2020.12.5
h11==0.9.0
hiredis==1.1.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
huey==2.3.0
idna==2.10
multidict==5.0.0
pip==20.3.3
python-dateutil==2.8.1
redis==3.5.3
rfc3986==1.4.0
sanic==20.9.1
sanic-validation==0.5.1
setuptools==49.2.1
six==1.15.0
sniffio==1.2.0
ujson==4.0.1
uvloop==0.14.0
websockets==8.1
